requirements.txt
fpdf2>=2.7.0
pillow>=9.0.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0